+++ lib/router.js
@@ -1,7 +1,7 @@
 import { Router } from "express";
-import { usersRouter, userRouter } from "./routes/user.js";
-import { inventoryRouter } from "./routes/inventory.js";
-import { authRouter } from "./routes/auth.js";
+import { usersRouter, userRouter } from "../routes/user.js";
+import { inventoryRouter } from "../routes/inventory.js";
+import { authRouter } from "../routes/auth.js";
 
 export const router = Router();
 router.get("/", (req, res) => {
