+++ models/Item.js
@@ -3,7 +3,7 @@ const { Schema } = mongoose;
 
 export const itemSchema = new Schema({
   name: { type: String, required: true },
-  user_id: { type: Schema.Types.ObjectId, required: true },
+  user_id: { type: Schema.Types.ObjectId, ref: "Item", required: true },
   description: String,
   image: String
 });
